geo-alcohol/requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
bonjour-py==0.3
cffi==1.2.0
cryptography==1.0
enum34==1.0.4
funcsigs==0.4
Genderize==0.1.4
google-api-python-client==1.4.1
httplib2==0.9.1
idna==2.0
ipaddress==1.0.14
macholib==1.5.1
matplotlib==1.4.3
mock==1.3.0
modulegraph==0.10.4
ndg-httpsclient==0.4.0
nltk==3.0.0
nose==1.3.7
numpy==1.9.2
oauth2client==1.4.12
pbr==1.5.0
py2app==0.7.3
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.15.1
pyparsing==2.0.3
python-dateutil==2.4.2
python-google-places==1.0.1
python-Levenshtein==0.12.0
pytz==2015.4
requests==2.7.0
rsa==3.2
Rtree==0.8.2
scikit-learn==0.16.0
scipy==0.14.1
simplejson==3.8.0
six==1.9.0
Twisted==13.2.0
uritemplate==0.6
virtualenv==13.1.0
wheel==0.24.0
xattr==0.6.4
zope.interface==4.1.1